refactor: 폴더 구조 개선
- features/applicants의 내용을 pages/applicants로 이동
- features/companies의 내용을 pages/myPage/employer로 이동
- features/recruitments의 내용을 pages/myCompany로 이동
