+++ src/pages/applicants/ApplicantList/ApplicantsTable.styles.ts
