+++ src/pages/applicants/ApplicantList/Contract/ModalText.styles.ts
