+++ src/pages/myCompany/Recruitments/RecruitmentInfo.styles.ts
