+++ src/pages/myCompany/Recruitments/RecruitmentList.tsx
@@ -1,23 +1,23 @@
 import { Flex, Typo } from '@/components/common';
-import CompaniesTable from './CompaniesTable/CompaniesTable';
-import { CompanyData } from '@/types';
+import RecruitmentsTable from './RecruitmentsTable';
+import { RecruitmentItem } from '@/types';
 
 type Props = {
-  companyList: CompanyData[];
+  recruitmentList: RecruitmentItem[];
 };
 
-export default function CompanyList({ companyList }: Props) {
+export default function RecruitmentList({ recruitmentList }: Props) {
   return (
     <Flex direction="column" justifyContent="center" gap={{ y: '15px' }}>
       <Flex alignItems="center" gap={{ x: '20px' }}>
         <Typo element="h3" size="24px" style={{ fontWeight: 'bold' }}>
-          내 회사
+          내 공고글
         </Typo>
         <Typo element="span" size="16px">
-          총 {companyList.length}곳
+          총 {recruitmentList.length}건
         </Typo>
       </Flex>
-      <CompaniesTable companyList={companyList} />
+      <RecruitmentsTable recruitmentList={recruitmentList} />
     </Flex>
   );
 }
