+++ src/pages/myPage/employer/Companies/CompanyInfo.styles.ts
