+++ src/pages/myPage/employer/Companies/CompanyList.tsx
@@ -1,23 +1,23 @@
 import { Flex, Typo } from '@/components/common';
-import RecruitmentsTable from './RecruitmentsTable/RecruitmentsTable';
-import { RecruitmentItem } from '@/types';
+import CompaniesTable from './CompaniesTable';
+import { CompanyData } from '@/types';
 
 type Props = {
-  recruitmentList: RecruitmentItem[];
+  companyList: CompanyData[];
 };
 
-export default function RecruitmentList({ recruitmentList }: Props) {
+export default function CompanyList({ companyList }: Props) {
   return (
     <Flex direction="column" justifyContent="center" gap={{ y: '15px' }}>
       <Flex alignItems="center" gap={{ x: '20px' }}>
         <Typo element="h3" size="24px" style={{ fontWeight: 'bold' }}>
-          내 공고글
+          내 회사
         </Typo>
         <Typo element="span" size="16px">
-          총 {recruitmentList.length}건
+          총 {companyList.length}곳
         </Typo>
       </Flex>
-      <RecruitmentsTable recruitmentList={recruitmentList} />
+      <CompaniesTable companyList={companyList} />
     </Flex>
   );
 }
